// tools/aimux-web/web/src/types/WireCallResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { WireMeta } from "./WireMeta";
import type { JsonValue } from "../../../../../bindings/node/src/types/serde_json/JsonValue";

/**
 * Response for non-streaming calls.
 */
export type WireCallResponse = { text: string, finish_reason: JsonValue, usage: JsonValue, meta?: WireMeta | null, error?: string | null, };